@@ -1,15 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 	  <RepositoryUrl>
 		  https://github.com/achiez/AchiesUtilities/
 	  </RepositoryUrl>
 	  <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-	  <Version>1.1.6</Version>
-	  <AssemblyVersion>1.1.6</AssemblyVersion>
+	  <Version>1.1.7</Version>
+	  <AssemblyVersion>1.1.7</AssemblyVersion>
   </PropertyGroup>
 
 
