@@ -1,11 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
-		<AssemblyVersion>1.0.6</AssemblyVersion>
-		<Version>1.0.6</Version>
+		<AssemblyVersion>1.0.7</AssemblyVersion>
+		<Version>1.0.7</Version>
 		<RepositoryUrl>
 			https://github.com/achiez/AchiesUtilities/
 		</RepositoryUrl>
@@ -13,15 +13,17 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<Folder Include="Proxy\" />
 		<Folder Include="Models\" />
 		<Folder Include="Builder\" />
 		<Folder Include="Extensions\" />
-		<Folder Include="Helpers\" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\AchiesUtilities\AchiesUtilities.csproj" />
 	</ItemGroup>
 
 </Project>
