@@ -1,6 +1,6 @@
 ﻿namespace AchiesUtilities.Web.Proxy;
 
-public enum ProxyType
+public enum ProxyProtocol
 {
-    HTTP, SOCKS4, SOCKS5
+    HTTP, HTTPS, SOCKS4, SOCKS5
 }
